+++ app/decorators/team_in_match_decorator.rb
@@ -1,0 +1,14 @@
+class TeamInMatchDecorator < SimpleDelegator
+  include Comparable
+
+  attr_reader :score
+
+  def initialize(team, score:)
+    @score = score
+    super(team)
+  end
+
+  def <=>(other)
+    (other.score || -1) <=> (@score || -1)
+  end
+end
